+++ pkg/crd/apply.go
@@ -31,24 +31,36 @@ import (
 
 	"github.com/arangodb/go-driver"
 
+	"github.com/arangodb/kube-arangodb/pkg/crd/crds"
 	"github.com/arangodb/kube-arangodb/pkg/logging"
 	"github.com/arangodb/kube-arangodb/pkg/util/kclient"
 )
 
 var logger = logging.Global().RegisterAndGetLogger("crd", logging.Info)
 
+// Deprecated: use EnsureCRDWithOptions instead
 func EnsureCRD(ctx context.Context, client kclient.Client, ignoreErrors bool) error {
+	return EnsureCRDWithOptions(ctx, client, nil, ignoreErrors)
+}
+
+func EnsureCRDWithOptions(ctx context.Context, client kclient.Client, ensureOpts map[string]crds.CRDOptions, ignoreErrors bool) error {
 	crdsLock.Lock()
 	defer crdsLock.Unlock()
 
-	for crd, spec := range registeredCRDs {
-		getAccess := verifyCRDAccess(ctx, client, crd, "get")
+	for crdName, crdReg := range registeredCRDs {
+		getAccess := verifyCRDAccess(ctx, client, crdName, "get")
 		if !getAccess.Allowed {
-			logger.Str("crd", crd).Info("Get Operations is not allowed. Continue")
+			logger.Str("crd", crdName).Info("Get Operations is not allowed. Continue")
 			continue
 		}
 
-		err := tryApplyCRD(ctx, client, crd, spec)
+		var opt *crds.CRDOptions
+		if o, ok := ensureOpts[crdName]; ok {
+			opt = &o
+		}
+		def := crdReg.getter(opt)
+
+		err := tryApplyCRD(ctx, client, def)
 		if !ignoreErrors && err != nil {
 			return err
 		}
@@ -56,45 +68,47 @@ func EnsureCRD(ctx context.Context, client kclient.Client, ignoreErrors bool) er
 	return nil
 }
 
-func tryApplyCRD(ctx context.Context, client kclient.Client, crd string, spec crd) error {
+func tryApplyCRD(ctx context.Context, client kclient.Client, def crds.Definition) error {
 	crdDefinitions := client.KubernetesExtensions().ApiextensionsV1().CustomResourceDefinitions()
 
-	c, err := crdDefinitions.Get(ctx, crd, meta.GetOptions{})
+	crdName := def.CRD.Name
+
+	c, err := crdDefinitions.Get(ctx, crdName, meta.GetOptions{})
 	if err != nil {
 		if !errors.IsNotFound(err) {
-			logger.Err(err).Str("crd", crd).Warn("Get Operations is not allowed due to error")
+			logger.Err(err).Str("crd", crdName).Warn("Get Operations is not allowed due to error")
 			return err
 		}
 
-		createAccess := verifyCRDAccess(ctx, client, crd, "create")
+		createAccess := verifyCRDAccess(ctx, client, crdName, "create")
 
 		if !createAccess.Allowed {
-			logger.Str("crd", crd).Info("Create Operations is not allowed but CRD is missing. Continue")
+			logger.Str("crd", crdName).Info("Create Operations is not allowed but CRD is missing. Continue")
 			return nil
 		}
 
 		c = &apiextensions.CustomResourceDefinition{
 			ObjectMeta: meta.ObjectMeta{
-				Name: crd,
+				Name: crdName,
 				Labels: map[string]string{
-					Version: string(spec.version),
+					Version: string(def.Version),
 				},
 			},
-			Spec: spec.spec,
+			Spec: def.CRD.Spec,
 		}
 
 		if _, err := crdDefinitions.Create(ctx, c, meta.CreateOptions{}); err != nil {
-			logger.Err(err).Str("crd", crd).Warn("Create Operations is not allowed due to error")
+			logger.Err(err).Str("crd", crdName).Warn("Create Operations is not allowed due to error")
 			return err
 		}
 
-		logger.Str("crd", crd).Info("CRD Created")
+		logger.Str("crd", crdName).Info("CRD Created")
 		return nil
 	}
 
-	updateAccess := verifyCRDAccess(ctx, client, crd, "update")
+	updateAccess := verifyCRDAccess(ctx, client, crdName, "update")
 	if !updateAccess.Allowed {
-		logger.Str("crd", crd).Info("Update Operations is not allowed. Continue")
+		logger.Str("crd", crdName).Info("Update Operations is not allowed. Continue")
 		return nil
 	}
 
@@ -104,21 +118,21 @@ func tryApplyCRD(ctx context.Context, client kclient.Client, crd string, spec cr
 
 	if v, ok := c.ObjectMeta.Labels[Version]; ok {
 		if v != "" {
-			if !isUpdateRequired(spec.version, driver.Version(v)) {
-				logger.Str("crd", crd).Info("CRD Update not required")
+			if !isUpdateRequired(def.Version, driver.Version(v)) {
+				logger.Str("crd", crdName).Info("CRD Update not required")
 				return nil
 			}
 		}
 	}
 
-	c.ObjectMeta.Labels[Version] = string(spec.version)
-	c.Spec = spec.spec
+	c.ObjectMeta.Labels[Version] = string(def.Version)
+	c.Spec = def.CRD.Spec
 
 	if _, err := crdDefinitions.Update(ctx, c, meta.UpdateOptions{}); err != nil {
-		logger.Err(err).Str("crd", crd).Warn("Create Operations is not allowed due to error")
+		logger.Err(err).Str("crd", crdName).Warn("Create Operations is not allowed due to error")
 		return err
 	}
-	logger.Str("crd", crd).Info("CRD Updated")
+	logger.Str("crd", crdName).Info("CRD Updated")
 	return nil
 }