+++ pkg/crd/arangomembers.go
@@ -25,5 +25,7 @@ import (
 )
 
 func init() {
-	registerCRDWithPanic(crds.DatabaseMemberDefinition())
+	registerCRDWithPanic(func(opts *crds.CRDOptions) crds.Definition {
+		return crds.DatabaseMemberDefinitionWithOptions()
+	}, nil)
 }