+++ pkg/crd/crds/apps-job.go
@@ -24,7 +24,6 @@ import (
 	_ "embed"
 
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	"k8s.io/apimachinery/pkg/util/yaml"
 
 	"github.com/arangodb/go-driver"
 )
@@ -34,23 +33,35 @@ const (
 )
 
 func init() {
-	if err := yaml.Unmarshal(appsJobs, &appsJobsCRD); err != nil {
-		panic(err)
-	}
+	mustLoadCRD(appsJobs, appsJobsSchemaRaw, &appsJobsCRD, &appsJobsCRDSchemas)
 }
 
+// Deprecated: use AppsJobWithOptions instead
 func AppsJob() *apiextensions.CustomResourceDefinition {
-	return appsJobsCRD.DeepCopy()
+	return AppsJobWithOptions()
 }
 
+func AppsJobWithOptions(opts ...func(*CRDOptions)) *apiextensions.CustomResourceDefinition {
+	return getCRD(appsJobsCRD, appsJobsCRDSchemas, opts...)
+}
+
+// Deprecated: use AppsJobDefinitionWithOptions instead
 func AppsJobDefinition() Definition {
+	return AppsJobDefinitionWithOptions()
+}
+
+func AppsJobDefinitionWithOptions(opts ...func(*CRDOptions)) Definition {
 	return Definition{
 		Version: AppsJobVersion,
-		CRD:     appsJobsCRD.DeepCopy(),
+		CRD:     AppsJobWithOptions(opts...),
 	}
 }
 
 var appsJobsCRD apiextensions.CustomResourceDefinition
+var appsJobsCRDSchemas crdSchemas
 
 //go:embed apps-job.yaml
 var appsJobs []byte
+
+//go:embed apps-job.schema.generated.yaml
+var appsJobsSchemaRaw []byte